pipe tag resolver to windows probe
