+++ pkg/security/resolvers/cgroup/model/model.go
@@ -9,8 +9,6 @@
 package model
 
 import (
-	"errors"
-	"fmt"
 	"sync"
 
 	"go.uber.org/atomic"
@@ -19,55 +17,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
 	"github.com/DataDog/datadog-agent/pkg/security/utils"
 )
-
-var (
-	ErrNoImageProvided = errors.New("no image name provided") // ErrNoImageProvided is returned when no image name is provided
-)
-
-// WorkloadSelector is a selector used to uniquely indentify the image of a workload
-type WorkloadSelector struct {
-	Image string
-	Tag   string
-}
-
-// NewWorkloadSelector returns an initialized instance of a WorkloadSelector
-func NewWorkloadSelector(image string, tag string) (WorkloadSelector, error) {
-	if image == "" {
-		return WorkloadSelector{}, ErrNoImageProvided
-	} else if tag == "" {
-		tag = "latest"
-	}
-	return WorkloadSelector{
-		Image: image,
-		Tag:   tag,
-	}, nil
-}
-
-// IsReady returns true if the selector is ready
-func (ws *WorkloadSelector) IsReady() bool {
-	return len(ws.Image) != 0
-}
-
-// Match returns true if the input selector matches the current selector
-func (ws *WorkloadSelector) Match(selector WorkloadSelector) bool {
-	if ws.Tag == "*" || selector.Tag == "*" {
-		return ws.Image == selector.Image
-	}
-	return ws.Image == selector.Image && ws.Tag == selector.Tag
-}
-
-// String returns a string representation of a workload selector
-func (ws WorkloadSelector) String() string {
-	return fmt.Sprintf("[image_name:%s image_tag:%s]", ws.Image, ws.Tag)
-}
-
-// ToTags returns a string array representation of a workload selector
-func (ws WorkloadSelector) ToTags() []string {
-	return []string{
-		"image_name:" + ws.Image,
-		"image_tag:" + ws.Tag,
-	}
-}
 
 // CacheEntry cgroup resolver cache entry
 type CacheEntry struct {
