+++ pkg/security/resolvers/opts_windows.go
@@ -1,0 +1,14 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+// Package resolvers holds resolvers related files
+package resolvers
+
+import "github.com/DataDog/datadog-agent/pkg/security/resolvers/tags"
+
+// Opts defines common options
+type Opts struct {
+	Tagger tags.Tagger
+}
